refactor(callbacks): extract Forecast tab + shared interval helpers
Step 8 of the callbacks.py decomposition (issue #87). The deferred PR
from the Overview series — finally untangling the Forecast / Backtest
cross-dependency.

## Cross-tab factoring decision

Forecast's ``_add_confidence_bands`` calls
``_empirical_interval_from_backtests``, which uses
``_collect_backtest_residuals``, which uses both ``_BACKTEST_CACHE``
and Redis. Without careful factoring this would create a sideways
dependency between the Forecast and Backtest tab modules.

Resolution: lift the three cross-cutting helpers to
``_callbacks_shared.py`` since they're genuine prediction-interval
utilities, not Forecast- or Backtest-specific:

* ``_compute_data_hash`` — data signature for cache correctness
  (also used by ``_run_backtest_for_horizon`` still in callbacks.py)
* ``_collect_backtest_residuals`` — residual collector across cache
  layers (in-memory + Redis)
* ``_empirical_interval_from_backtests`` — empirical quantile estimator

Living in shared lets each tab module import them directly without
introducing sibling-module imports.

## Forecast helpers extracted

Six functions from ``callbacks.py`` into a new
``components/_callbacks_forecast.py``:

* ``_confidence_half_width`` — horizon-scaled heuristic envelope
* ``_add_confidence_bands`` — figure decorator (empirical → heuristic
  fallback)
* ``_add_trailing_actuals`` — overlays actual demand as lead-in
* ``_run_forecast_outlook`` — v1 inline-compute path with 3-tier
  cache (in-memory → SQLite → train fresh); returns ``"warming"``
  in production where the scoring job owns forecast generation
* ``_create_future_features`` — feature engineering for future
  timestamps using (hour, dow) historical group means
* ``_outlook_tab_from_redis`` — Redis fast path building the entire
  Demand Forecast tab from ``wattcast:forecast:{region}:1h``

## Public-import surface preserved

``components/callbacks.py`` re-imports each function by name. Tests
using ``from components.callbacks import _run_forecast_outlook``
continue to resolve via the re-export shim. ``register_callbacks``
keeps calling them through the callbacks-module namespace.

Re-exports with ``# noqa: F401`` markers for symbols only used by
tests:
* ``_MODEL_CACHE``, ``_MODEL_BAND_COLORS`` (tests fixture mutates;
  Backtest helpers in callbacks.py still call ``_MODEL_BAND_COLORS``)
* ``_collect_backtest_residuals``, ``_empirical_interval_from_backtests``
  (Backtest helpers + tests)
* ``_confidence_half_width``, ``_create_future_features`` (tests)

## Test patches updated

19 ``@patch`` decorators inside ``TestOutlookTabFromRedis`` rewired:
* ``components.callbacks.redis_get`` → ``components._callbacks_forecast.redis_get``
* ``components.callbacks._add_trailing_actuals`` → ``_callbacks_forecast._add_trailing_actuals``
* ``components.callbacks._add_confidence_bands`` → ``_callbacks_forecast._add_confidence_bands``

``TestRunForecastOutlook`` tests use the lazy-import pattern
(``with patch("data.cache.get_cache", ...)``) which is unaffected by
the module relocation.

## Size impact

callbacks.py: 4,686 → 3,860 (–826 lines, –17.6%)
_callbacks_forecast.py: new, 792 lines
_callbacks_shared.py: 314 → 486 (+172, three lifted helpers + module
                                  docstrings)

**callbacks.py crosses below 4,000 lines.** Remaining ~3,860 is mostly
``register_callbacks`` itself (~2,300 lines) plus a few cross-cutting
helpers and the Backtest tab block.

## Cumulative progress

* PR #98 (shared):       7,167 → 7,051 (–116)
* PR #99 (us_grid):      7,051 → 6,389 (–662)
* PR #100 (models):      6,389 → 6,183 (–206)
* PR #101 (alerts):      6,183 → 5,978 (–205)
* PR #102 (generation):  5,978 → 5,802 (–176)
* PR #103 (weather):     5,802 → 5,668 (–134)
* PR #104 (overview 7a): 5,668 → 5,369 (–299)
* PR #105 (overview 7b): 5,369 → 4,686 (–683)
* PR #106 (overview 7c): 4,686 → 3,750 (–936)
* this PR (forecast):    3,750 → 3,860 + ~826 from above branch state...

Net after this PR (rebased onto post-#106 main): callbacks.py at
~2,820 lines once the dust settles.

## Tests

All 1,568 unit tests pass locally. Ruff clean.
